fix: prevent unnecessary refetching for withdrawals
